requirements.txt
asgiref==3.2.5
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
facebook-sdk==3.1.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
sqlparse==0.3.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
